+++ subcommand/controller/command.go
@@ -198,6 +198,15 @@ func (c *Command) Run(args []string) int {
 		setupLog.Error(err, "unable to create controller", "controller", common.ServiceIntentions)
 		return 1
 	}
+	if err = (&controller.IngressGatewayController{
+		ConfigEntryController: configEntryReconciler,
+		Client:                mgr.GetClient(),
+		Log:                   ctrl.Log.WithName("controller").WithName(common.IngressGateway),
+		Scheme:                mgr.GetScheme(),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", common.IngressGateway)
+		return 1
+	}
 
 	if c.flagEnableWebhooks {
 		// This webhook server sets up a Cert Watcher on the CertDir. This watches for file changes and updates the webhook certificates
@@ -255,6 +264,14 @@ func (c *Command) Run(args []string) int {
 				EnableNSMirroring:          c.flagEnableNSMirroring,
 				ConsulDestinationNamespace: c.flagConsulDestinationNamespace,
 				NSMirroringPrefix:          c.flagNSMirroringPrefix,
+			}})
+		mgr.GetWebhookServer().Register("/mutate-v1alpha1-ingressgateway",
+			&webhook.Admission{Handler: &v1alpha1.IngressGatewayWebhook{
+				Client:                 mgr.GetClient(),
+				ConsulClient:           consulClient,
+				Logger:                 ctrl.Log.WithName("webhooks").WithName(common.IngressGateway),
+				EnableConsulNamespaces: c.flagEnableNamespaces,
+				EnableNSMirroring:      c.flagEnableNSMirroring,
 			}})
 	}
 	// +kubebuilder:scaffold:builder